requirements.txt
asgiref==3.4.1
boto3==1.20.24
botocore==1.23.24
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.10.0
Django==4.0
djangorestframework==3.13.1
idna==3.3
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
mysql-connector-python==8.0.27
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
s3transfer==0.5.0
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tzdata==2021.5
urllib3==1.26.7
